Regroup files
