+++ Sources/VGSCollectSDK/Core/API/APIHostName/APIHostnameValidator.swift
